requirements.txt
aiohttp==3.4.4
alabaster==0.7.12
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
docutils==0.14
idna==2.7
imagesize==1.1.0
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.3.0
multidict==4.4.2
packaging==18.0
pluggy==0.8.0
py==1.7.0
Pygments==2.2.0
pyparsing==2.2.2
pytest==3.9.2
python-dateutil==2.7.4
pytz==2018.6
requests==2.20.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.12
urllib3==1.24
yarl==1.2.6
numpy==1.15.2
numpydoc==0.8.0
pandas==0.23.4
pandas-datareader==0.6.0
